backend/requirements.txt
Django==5.1.6
psycopg2-binary==2.9.9
gunicorn==22.0.0
django-environ==0.11.2
django-cors-headers==4.4.0
fal-client>=0.6.0
Pillow==11.1.0
